Fixed Issue #470
I am uncertain wether the key should be set on input settings screen.
The issue is fixed but I believe there are several key events which
shouldn't be handled in the DebugControlSystem class - This being one of
them. Then again I am not that familiar with the source. I just don't
believe removing the hud should be a debug thing.