Updated printScreen method to lookup bpp value.
The bpp displaymode query returns 0 in windowed mode, or actuall bpp
in fullscreen mode. If the 0 value is detected in non-fullscreen mode, the
logic queries the display for the DesktopDisplayMode.  Finally if value is
still at 0 bpp, overrides with 32 (equivalent to original 4 value that was
hardcoded for RGBA).  Updated values relying on the bpp value to take into
account actual byte count value from bits by dividing by 8.