Replace entity refs when creating component copies.
The entites get replaced by a new EntityRef implementation called
DelayedEntityRef. This new EntityRef implementation will obtain
and delegate to an actual EntityRef once it gets accessed for the first
time. The DelayedEntiryRef can be created on the main thread and
used on the saving thread with it's private entity manager.