Check if a mesh is hidden during rendering to handle owner changes.
Reasoning: The game no longer saves client entities. Thus when a player
gets loaded it is not attached to a client. Thus it is not possible
to check at the creation time of the player entity if it's mesh is
visible as that check requires the link to the client entity that is
not there yet. Thus the check for the meshes visiblity is now done
during rendering.