Remove the placeBlock command as it does not work in multiplayer.
Reasoning: The command is not worth fixing for multiplayer as it is
useless. Blocks can be placed via right click. The bad thing about the
command is that it appears to be working but in multiplayer too, but
the blocks exist only on client side only.

Interesting for the future might be a command for placing multiple
blocks at once at a certain location but as it is the command is
useless.