Renamed LwjglRenderingProcess to FrameBuffersManager.
This is due to previous changes having stripped the LwjglRenderingProcess class of most of its functionality (now available through the GraphicState and PostProcessor classes). What's left is only FBOs-related functionality, hence the new name. This commits introduces no functional changes.  Most changes in this commit are a consequence of the renaming. The only exception is the class-level javadoc for the FrameBuffersManager class, which has been converted into a TODO.