Give NUIManager a better fitting life cycle and context instance.
The NUIManager is no longer in the engine context.

Where previously it just got "refreshed" now a new instance gets created.

Namely the first NUIManager like the first sub context for the duration
of the main menu.

When the starts loading a new NUIManager gets created that lifes again
like it's owning Context till the end of the game.

This patch fixes the issue that caused interations require two activations
when a window got closed with escape.