Overworked how view distances get changed.
Previously it was broken because a binding expected a standard setter
for the view distance and there was none.

The new approach introduces a ViewDistanceChangeRequest evetn that gets
sent and processed by the client. It triggers a  ViewDistanceChangeEvent
that informs the server about the change, so that it knows which
chunks to send.