Add distinguished select game screen
Add distinguished select game screen for singleplayer and host game.
Singleplayer now has a subtitle "singleplayer" and host game has a
subtitle "multiplayer".
Created a SelectGameScreenHost class (copy of SelectGameScreen) just to
be able to couple the right .ui file to it (no idea if this is the right
way to make the distinction)