Fix a memory leak by properly clearing recorded data about temporary blocks.
The EntitySetDeltaRecorder class records changes to persistent entities
in order to forward them later on to the background saving thread.

It got informed about changes to temporary block entities, but did not
get informed about the silent destruction of those entities when the
entities were no longer needed to represent the block.

This patch marks the entity manager inform the EntitySetDeltaRecorder
indirectly about the destruction of the temporary block entities,
so that it knows that the entities do no longer need to be stored.