Fixes issue #2452, tooltips flicker when forced under the mouse
This scenario can occur when a tooltip is unable to render in the preferred
location due to the screen boundaries, for instance if an item with a really
long tooltip is placed in the lower-right most inventory slot.  Because UILists
define InteractionRegions around their contents by default and tooltips are
rendered last, they are seen as the new top-most interactive element on the
following game tick.  This causes an immediate reset of the tooltip timer due
to the top-most region changing, and the process repeats.  This has been fixed
by preventing tooltips from defining interaction regions through a new optional
flag in UIList that sets whether the list is interactive or not.