Fix first letter in chatbox moving player
Fix part of issue #2833 (Various isses with movement inputs persisting
through the chatbox) where the first letter typed into an empty chatbox was
causing the player to move.

This was caused by the event falling through to the normal event system,
because a path was not marked with eventHandled = true.