Fix give command only giving 1 item at a time
Fixes the main issue in #2835. The command to give multiple items was just
giving one at a time, but now it will try to give the actual number entered.

If the number entered causes the inventory to overflow, no more items will
be given, and the actual number of items given will be output in the command
window.