Make give command give 16 stackable blocks, or 1 non-stackable block
As mentioned in #2835, when the give command is called without a quantity,
it should give 16 blocks if the block is stackable, or 1 block if not. The
command should correctly report the number of blocks given.

This was temporarily broken by 6234a3, because the previous implementation
tried to give a stack of 16, but that was automatically corrected to 1 (the
command reported that 16 blocks were given).

The command now gives the correct number of blocks, and reports the actual
number given.