Warning if running in 32-bit JVM on 64-bit system
If the game is running in a 32-bit JVM on a 64-bit system, a warning is shown in the console and on the main menu screen.
This can help users getting limited performance due to running in a 32-bit JVM when they could install a 64-bit one (thus native for their CPU), like described in issue #2905.
On Windows, this is checked by simply verifying if the PROCESSOR_ARCHITEW6432 is set.
On POSIX OSs, the JVM architecture is read from the "os.arch" property and the real system architecture by reading the output of "uname -m".