Final fix for now: Use a less performant loop approach to catching non-triggered held items.
This works for two scenarios:
* When logging into a game for the first time any existing remote players will be hooked up with held item entities
* When another play logs in any existing player will correctly get a new held item entity for their item

In both cases this only matters if the players are actually holding visible things. Even without the fix after a player changes their item once it shows up for all.

Two separate one-time approaches (a postBegin and a player join event) would be superior for performance reasons.