Changed InventoryScreen.class so movingItem not lost
Fixes an issue where the item in the transfer slot would stay in the transfer slot when the inventory was closed. It would show up on the cursor in the pause screen and even the main menu. It would only disappear when the save was loaded, at which point the item would be lost.

Now what happens is that if an item is in the transfer slot when the inventory screen is closed, the item will be placed in the first avaliable inventory slot. If there are no avaliable inventory slots the item will be dropped in front of the player.

This has been done by creating an override for the InventoryScreen's onClose() method and cannibalizing some of the code in the InventoryCell and DropItemRegion classes which had similar functionality for different applications.

(This only has one commit because I was working on this locally while I messed around trying to figure out how the game works and ended up finding the solution before I forked the repo.)