postAutoSave() now takes item back out of correct slot in inventory
Currently only preAutoSave() and postAutoSave() methods. Will add preSave() and postSave() when everything is working properly.

This commit adds code to postAutoSave() and a variable defined at the top of the class to get the correct item back out of the player's inventory after the save is complete. It also adds an if statement around the bulk of the code in preAutoSave() to make sure there actually is an item in the transfer slot before doing anything.

Right now if there is a stack of items in the transfer slot and the items are sorted into multiple slots, only the items that ended up in the right-most slot will be put back into the transfer slot.

There is also nothing checking whether the screen is open or not. (Though as far as I can tell that isn't causing any immediate problems.)

The println statements will be removed once this is fully functional.