The game details screen now shows the world name ("Perlin-0"/"Flat-1"/etc.) instead of "main".
WorldInfo() objects are created with another String member variable "customTitle" which stores the world name. The "title" variable is preserved as currently the StoragePathProvider system uses "main" as the folder name for the main world.

In GameDetailsScreen.setUpGameWorlds(), the gameWorlds UIList uses setItemRenderer() to draw the "customTitle" String variable on the UIList widgets.