Add ShortBuffer and FloatBuffer to API whitelist
These are useful for audio buffers. While it's still necessary to
copy to a `ByteBuffer` when passing it on to Terasology, this
helps a lot during synthesis and effect processing.

Additionally, these are analogous to `ByteBuffer` and `IntBuffer`. So
it seems reasonable to have these too. (And in a future PR,
maybe `DoubleBuffer`, `LongBuffer` and `CharBuffer`?)