Fix for Z-fighting on back faces of blocks
This prevents the interior faces of partially translucent blocks from being merged into the exterior faces of the blocks behind them.

It would be wise to only adjust interior faces, and not exterior faces, but as of yet I have been unable to determine the proper way to accomplish this.