Added controller button rebinding and axis sensitivity setting
It is now possible to rebind controller buttons. This is done via
sending a ControllerButtonEvent that is then handled by NuiManagerInternal
and UIInputBind.

The controller settings page now contains an axis sensitivity assigned to every axis.
The name labels were previously the raw names from the controller, but are now converted to a display name.

One controller event test was added to InputSystemTest.

Changes were tested with an XBOX 360 controller on Linux.