Revert "Restore pre-BiomeAPI foliage shading (#3802)"
This reverts commit 8ad3d164

The method in which I had implemented this breaks the server/client boundary by asking for server-specific information about the world in the client. It works fine in single-player, but of course it causes crashes when attempting to connect to servers. This feature still needs to be reimplemented, but it will be a much larger job that requires more infrastructure for obtaining the required information from the server.