chore(groovyw): Add command to list local modules
Adds a flag '--local' that can be added to `groovyw module list` in
order to only list modules that are present locally as source.

The output of this command is machine readable such that it can be used
easily to perform some command for all modules.

For instance, the following Bash command will reset and update all
modules:

```sh
for m in $(./groovyw module list --local)
do
    pushd modules/$m
    git reset --hard HEAD
    git pull
    popd
done
```