Make deep copying optional
Add an argument to CopyStrategy for whether to deep copy entities,
so that the original behaviour is still available. Entities are
copied deeply only if they are owned by the object being copied
and the deep-copy version of the copy function is called.
EntityRef.copy does a deep copy by default.