slab stacking, block family refactor
The main intention of this commit is to permit the horizontal block
family to stack correctly.  In other voxel worlds, half blocks can
be replaced in the voxel with a full block of a similar type (wood
slab with full wood block).  This commit does not attempt to handle
placing blocks on the underside since that requires a little more
work on the physics and targeting systems for blocks.

One observation about the state of the situation is that a lot of
the block placement decision making is done by BlockItemSystem
instead of the block families.  This commit takes the position that
these decisions should be parameterized by the target block family
and the placing block family.

While refactoring to move some of this power to BlockFamily, some
of the previous deprecated or vestigial constructors and methods
were also cleaned up.  Some of this looked to be unreachable
code.

This commit was tested by modifying the core assets library to have
CoreAssets:HalfPlank with a fullBlock := CoreAssets:Plank.