Add `shapes` attribute to block files, and remove FreeformFamily
Re-enable multiple blocks per shape in a more controlled manner,
requiring all of them to be registered in the block definition.
Blocks without a shape now default to cube rather than freeform.
Existing block files should be usable unchanged except in the case
where the freeform feature was actually used, in which case the list
of possible shapes will have to be added.