Add ElevationFacet and SurfacesFacet to replace SurfaceHeightFacet.
Split the SurfaceHeightFacet in two, with each part to be used for
somewhat different purposes, to improve the flexibility of the
world generator, particularly regarding multiple surfaces in the
same vertical column (e.g. overhangs). Leave SurfaceHeightFacet
present but deprecated, so that modules don't have to all transition
at once.