fix(DiscordRPC): NPE on shutdown
This was causing a NPE exception (not sure whether on `player` or `delayManager`) on shutdown, causing the game process to stall and the game did not exit properly.

If either of `player` or `delayManager` are no longer present we just don't do anything about it and continue with the other shutdown actions.