feat: add priority to Updates annotation of facet providers (#4833)
`WorldBuilder` comes up with the order to run facet providers in based on their annotations - `@Requires`, `@Produces`, and `@Updates`. Currently, there's no way to control order between `@Updates` providers - for example, a river provider should update the elevation after a mountain provider, so that rivers are carved into mountains and not the other way around. 

This PR adds a `priority` field with similar effect to the existing event handler priority mechanism to `@Updates`. This also ended up simplifying the provider ordering code.