feat: make render distance for (partially) transparent blocks adjustable (#4858)
Transparent blocks like leaves and grass are usually invisible beyond a certain distance. This is actually because the game will only render transparent blocks in the closest 64 chunks to the player. 

This PR disables the limit by default, but it can be re-enabled and the limit changed with a slider in the video settings.

* fix: render (partially) transparent blocks at any distance
* feat: add a video setting for billboard limit
* chore: clean up in response to checkstyle and intellij lints