bugfix: PojoPrefab was getting seriliazed as a value type (#4862)
PojoPrefab was getting serialized as a value type when wrapped in `RuntimeDelegateTypeHandler` due to a conflict between `AssetTypeHandler` and `PrefabTypeHandler`. We end up registering `Prefab` twice; once from evaluation and another from a factory. 

This caused the prefab to be incorrectly serialized from the server to the client. 
There is a cache in which we keep track of the types so its getting resolved to `AssetType` on the server when serializing and resolved to `PrefabType` on the client side.

Removing the `PrefabTypeHandler` should resolve this conflict.