feat: add convenience method  PhyiscEngine#recomputeCharacterCollider(entity)
We have code like this in some place, seemingly to make sure the character collider is recomputed to reflect changed values on the entity's components.

```java
physicsEngine.removeCharacterCollider(entity);
physicsEngine.getCharacterCollider(entity);
```

With the ability to simply add this as default method implementation in the interface we can provide a documented implementation for this, introducing a better alternative to cryptically calling `remove...` and `get...`.