fix(subsystems/discord): catch runtime exception and log only message
I noticed full stracktraces being logged from the subsystem for Discord
due to missing sockets or files. These come from runtime exceptions that
we don't catch.

I think that the respective error message is sufficient (in most cases)
to inform that the Discord subsystem could not be loaded. This reduces
the log spam during startup.