ci: fix project-autoadd github action (#5080)
A fix for the [`project-autoadd` github action](https://github.com/MovingBlocks/Terasology/blob/develop/.github/workflows/project-autoadd.yml) which failed because the utilized underlying action only supports classic projects but we're using v2 (beta) projects. The replacement action not only works but is a base action provided by GitHub which hopefully means better maintenance and higher security awareness.

The action requires a token with control privileges for projects. I added such a token to the organization secrets and wrote a [maintainer guide](https://github.com/MovingBlocks/Terasology/wiki/Maintenance#how-to-fix-an-expired-github-action-token) on what to do in case the action starts failing because this token expired.