fix: remove Rendering Module Settings screen
We currently don't have a model that allows us to easily load (parts) of modules to make them more configurable in the game setup screens.

In particular, the `RenderingModuleRegistry` is not yet initialized, and thus the whole purpose of the rendering module configuration is defied.

Therefore, we remove the screen from the menu and also remove the (now dead) code from the repo.

Related to #5076.