typehandlerlibrary qa
as a workaround, add dependency for spotbugs annotations, and create a
[ticket there so the spotbugs gradle plugin would include the dependency]
(https://github.com/spotbugs/spotbugs-gradle-plugin/issues/1018).

see here a motivation for inner type last rule. if we leave the rule we
should respect it, otherwise disable the rule.
https://stackoverflow.com/questions/25158939/what-motivation-is-behind-checkstyle-inner-type-last-rule

especially in tests just quite the warnings, in most cases the code is
deliberate as it is.

see #3859 and see bugs:
* https://github.com/pmd/pmd/issues/4731
* https://github.com/spotbugs/spotbugs-gradle-plugin/issues/1018
* https://github.com/spotbugs/spotbugs/issues/2667