refactor: allow-deny lists refactor also allows for backwards compatibility
Code strucutre is different in SelectGameScreen and ServerConnectListManager
due to compiling error otherwise.
We check if path exist then we set the value if it does,
otherwise it applies "previous" behaviour (meaning in case if even blacklist or whitelist)
do not exist it sets the file to null.

So no different behaviour is to be expected.