permit install into windows program files
when installing terasology into program files, that dir is not writeable.
starting up, terasology tries to create a folder in .local, which then
fails.

therefor, permit to start the application from a different directory,
and let it anyway find its libraries via the classpath. a workaround,
as the app should ideally know its location and other jarfiles without
tinkering with the classpath.

Co-authored-by: BenjaminAmos <24301287+BenjaminAmos@users.noreply.github.com>